Restructured doc generation service
