+++ server/controllers/proximityAlert/crimeVersion.ts
@@ -10,7 +10,7 @@ import presentCrimeVersion from '../../presenters/crimeVersion'
 import CrimeService from '../../services/crimeService'
 import MapImageRendererService from '../../services/proximityAlert/proximityAlertMapImageService'
 import PlaywrightBrowserService from '../../services/proximityAlert/playwrightBrowserService'
-import ProximityAlertReportDocxService from '../../services/proximityAlert/proximityAlertReportDocxService'
+import ProximityAlertReportDocxService from '../../services/proximityAlert/reportDocx/proximityAlertReportDocxService'
 import presentProximityAlertReportData from '../../presenters/proximityAlertReportData'
 import type { MojAlert } from '../../types/govUk/mojAlert'
 import { createMojAlertWarning } from '../../utils/alerts'
