+++ server/services/proximityAlert/reportDocx/proximityAlertReportDocxService.test.ts
@@ -1,6 +1,15 @@
+import dayjs from 'dayjs'
+import utc from 'dayjs/plugin/utc'
+import timezone from 'dayjs/plugin/timezone'
+import customParseFormat from 'dayjs/plugin/customParseFormat'
+
 import ProximityAlertReportDocxService from './proximityAlertReportDocxService'
-import type { ProximityAlertReportData } from '../../presenters/proximityAlertReportData'
-import type { ProximityAlertReportImages } from './proximityAlertMapImageService'
+import type { ProximityAlertReportData } from '../../../presenters/proximityAlertReportData'
+import type { ProximityAlertReportImages } from '../proximityAlertMapImageService'
+
+dayjs.extend(utc)
+dayjs.extend(timezone)
+dayjs.extend(customParseFormat)
 
 // Minimal valid 1x1 JPEG
 const tinyJpeg = Buffer.from(
