+++ server/services/proximityAlert/reportDocx/sections/exhibitPositionsSection.ts
@@ -1,0 +1,116 @@
+import { AlignmentType, Paragraph, ShadingType, Table, TableCell, TableLayoutType, TextRun, WidthType } from 'docx'
+import type { ProximityAlertReportDeviceWearer } from '../../../../presenters/proximityAlertReportData'
+import { formatDateTime } from '../../../../utils/date'
+import {
+  defaultCellProps,
+  cellParagraph,
+  defaultHeaderCellProps,
+  rowNoSplitAcrossPages,
+  sectionHeaderShading,
+  strongBlackBorders,
+} from '../docxComponents'
+
+const fmtDateTime = (dateString: string): string => formatDateTime(dateString, 'DD/MM/YYYY HH:mm')
+
+const exhibitPositionsSection = (wearer: ProximityAlertReportDeviceWearer): Table => {
+  const borders = strongBlackBorders()
+
+  const headingRow = rowNoSplitAcrossPages([
+    new TableCell({
+      ...defaultCellProps(),
+      borders,
+      shading: sectionHeaderShading(),
+      columnSpan: 7,
+      children: [
+        cellParagraph(`Exhibit EMAC/03 - Table of ${wearer.name}'s locations within the vicinity`, {
+          bold: true,
+          alignment: AlignmentType.CENTER,
+        }),
+      ],
+    }),
+  ])
+
+  const descriptionRow = rowNoSplitAcrossPages([
+    new TableCell({
+      ...defaultCellProps(),
+      borders,
+      columnSpan: 7,
+      children: [
+        new Paragraph({
+          children: [
+            new TextRun(
+              'The below table displays the location points within the crime vicinity. Please note the sequence number is ordered chronologically based on locations within the vicinity. ',
+            ),
+            new TextRun({
+              text: "Note: Consecutive numbers do not necessarily indicate sequential movement - please review the time stamps of each point to understand the offender's movements.",
+              italics: true,
+              bold: true,
+            }),
+          ],
+          spacing: { before: 0, after: 0 },
+        }),
+      ],
+    }),
+  ])
+
+  const headerCell = (text: string, widthPct: number): TableCell =>
+    new TableCell({
+      ...defaultHeaderCellProps(),
+      borders,
+      shading: { type: ShadingType.CLEAR, color: 'auto', fill: '4472C4' },
+      width: { size: widthPct, type: WidthType.PERCENTAGE },
+      children: [
+        cellParagraph(text, {
+          bold: true,
+          color: 'FFFFFF',
+          alignment: AlignmentType.CENTER,
+          size: 17,
+          spacingBefore: 5,
+          spacingAfter: 5,
+        }),
+      ],
+    })
+
+  const dataCell = (text: string, shaded: boolean): TableCell =>
+    new TableCell({
+      ...defaultCellProps(),
+      borders,
+      shading: shaded ? { type: ShadingType.CLEAR, color: 'auto', fill: 'D9E2F3' } : undefined,
+      children: [cellParagraph(text, { alignment: AlignmentType.CENTER })],
+    })
+
+  const headerRow = rowNoSplitAcrossPages([
+    headerCell('SEQUENCE NO.', 13),
+    headerCell('DATE/TIME', 15),
+    headerCell('LATITUDE\n(WGS84)', 13),
+    headerCell('LONGITUDE\n(WGS84)', 14),
+    headerCell('CONFIDENCE\nCIRCLE (Radius - m)', 18),
+    headerCell('SPEED (km/h)', 13),
+    headerCell('DIRECTION\n(degrees)', 14),
+  ])
+
+  const dataRows = wearer.positions.map((position, index) => {
+    const shaded = index % 2 === 1
+    const speed = 'speed' in position ? String(position.speed) : ''
+    const direction = 'direction' in position ? String(position.direction) : ''
+
+    return rowNoSplitAcrossPages([
+      dataCell(position.sequenceLabel, shaded),
+      dataCell(fmtDateTime(position.capturedDateTime), shaded),
+      dataCell(String(position.latitude), shaded),
+      dataCell(String(position.longitude), shaded),
+      dataCell(String(position.precision), shaded),
+      dataCell(speed, shaded),
+      dataCell(direction, shaded),
+    ])
+  })
+
+  return new Table({
+    width: { size: 100, type: WidthType.PERCENTAGE },
+    layout: TableLayoutType.FIXED,
+    borders,
+    rows: [headingRow, descriptionRow, headerRow, ...dataRows],
+  })
+}
+
+export default exhibitPositionsSection
